Inline AbstractChangesetVersionCommand.
There was only one subclass left, so it didn't make sense to keep the
extra abstraction.
